+++ config.yml
@@ -1,0 +1,5 @@
+# For the sensory detector
+directories_to_check:
+  - /home/omar2535/school/Rescate-tance/sensory-detector
+  - /home/omar2535/school/Rescate-tance/
+frequency_to_check: 1 # in seconds
