+++ detector/Detector.py
@@ -1,0 +1,10 @@
+from general_utils.CustomLogger import CustomLogger
+from general_utils.log_utils import do_rollover
+
+# General detector class
+class Detector:
+    def __init__(self):
+        pass
+
+    def run(self):
+        raise Exception("Overwrite me!")
