+++ main.py
@@ -1,0 +1,4 @@
+from sensory_detector.SensoryDetector import SensoryDetector
+
+detector = SensoryDetector()
+detector.run()
