+++ sensory_detector/config.yml
