+++ sensory_detector/logs/.gitignore
